Added image screensaver
Necessary to take a screenshot of a web page and save it as an image for further processing to be eventually displayed on a device.

Reek has been updated to ignore, for now, the feature envy and too many statements code smells because breaking out the logic for capturing an image doesn't make sense to extract further.