Fixed YJIT detection
Necessary to ensure YJIT is only enabled if YJIT is present since some people will install Ruby without YJIT support.

Code coverage is temporarily disabled because there isn't a quick way to test the `else` branch. One solution would be to move this to an initializer but want to hold off on doing that and return to this in the future.

Issue: 1
Milestone: patch