Updated to Hanami View 2.2.1
Necessary to finally leverage multiple scopes. This has been an oustanding bug that took nearly a year and a half to resolve so is nice to no longer have to maintain our provide our own patch for this.

This includes patches for Dry Monads and RuboCop AST as well.

Milestone: patch