Refactored models as aspects
These don't need to be at the top level of the app and can be namespaced within `Aspects` since they are ancillary objects which aid in API responses. This also prevents these objects from being auto loaded so the pragma for each file is no longer needed as well.

Milestone: patch