Fixed API display show action fetcher injection
Necessary to ensure the proper namespace is used since this was causing a stack dump. This popped up due to the recent namespace refactor and didn't manage to catch this properly. I've added an additional spec to ensure the fetcher dependencies is properly called so this won't happen again.

Milestone: patch