Added API log payload processing
Necessary to capture log information for devices in order to improve the debugability of individual devices.

This replaces the original HTTP header dump to the logs in favor of logging what we get back from the device straight to the database. Each log is associated with a device. This information will eventually be used for display within the UI.

Milestone: minor