Added screen editor middleware
Allows us to produce Server Side Events (SSE) using Rack middleware since Hanami actions don't have the ability to stream data even though they inherit from `Rack::Response`. The newly added `EventStream` is used to stream data to the client.

This is meant to be initialized with a pattern (i.e. route path). The pattern is used to pattern matching on route path and also obtain the ID which is sent to the event stream to create a unique image file for rendering purposes. This allows this object to be registered for different routes if desired.

Issue: 5
Milestone: minor