Fixed screen fetcher and new device setup template to use SVG image
Necessary to use consistent images when creating a virtual device and wanting to see the same image for both the new and show templates. This change reduces that confusion.

This also renames the file to use `setup.svg` which aligns with `setup.bmp`. The former is used within the UI for new device setup and the latter is used by the device when hitting the Setup API.

Milestone: patch