Added device logs migration for max alloc size
This information is provided in Firmware 1.5.2 so we can capture this information for logging purposes.

There is additional information, being provided, such as `filename_new` and `filename_current` that shows up in the request payload but they are blank and appear to be not completely thought through so am ignoring until this stabilizes further.

Milestone: minor