Updated pollers to be more customizable when called
This came up in a Discord discussion where a person wanted to use a different default for seconds and didn't realize you could update the associated script that calls these pollers. By allowing seconds to be customized when called, instead of when initialized, will make this more convenient for folks.

Milestone: minor