Removed device fields popover text for firmware bug
This is still a firmware bug but this is fixed in this app because we only provide the firmware URL if the versions don't match. Otherwise, the firmware URL is `nil` and prevents the device updating in an endless loop. So you can keep this enabled, as desired, without having to worry about manually toggling it on an off!

Milestone: patch