Updated screen downloader to use deduplicator
Necessary to prevent the proliferation of duplicate images now that we can inject and use the deduplicator. This also renames all files as `proxy-<id>.<extension>` which makes them shorter to read and clearly identifies them as proxy images that come from the remote server.

We no longer need to use MiniMagick to determine type since we can obtain file name and type by parsing the URI.

Milestone: minor