Added screen types
Necessary to define the supported image types. This'll soon be used by several objects which is why this is a separate file and constant. We could also make this an application setting but don't want to open up customization as of yet because we are beholden to what the device supports.

Zeitwerk has been taught to recognize the "types" file as the "TYPES" acronym otherwise this constant won't load properly.

Milestone: minor