Fixed API screens create action code coverage
Necessary to ensure we maintain code coverage when saving images by not using pattern matching but checking if the result is a success instead. Still not ideal but better than before. Reek is detecting feature envy issues with `result` but there isn't better way to reduce this further so am ignoring for now.

Milestone: patch