Added problem details application configuration format type
Necessary to register this type for API error responses. This will allow actions to use the `:problem_details` type instead of the default `:json` type when the situation calls for it.

Unfortunately, due to a Hanami bug, all default types must be re-registered when a custom type is added. 💢

Milestone: minor