Added view time at helper
Necessary to render time information in a format usable by HTML input elements with `time` type and also for human readability within the UI. This'll soon be used by device templates and partials.

This also refactores `.human_at` to be consistent with existing terminology within this module. Plus the double use of `at` was distracting.

Issue: 109
Milestone: minor