Fixed device new action and view field defaults
Necessary to ensure defaults are supplied to the view via the controller. The mistake was thinking you can use a block for the defaults in the view but a block only lazy loads values as actual values (not defaults). This prevents the issue of erasing manually entered values when the form has errors.

Milestone: minor