Updated screen fetcher to accept device
Necessary because we need to start obtaining sleep information from devices in order to determine if we should render a sleep image or not. These changes allow the device to be passed in instead of the device slug so we can ask the device for more information.

Issue: 109
Milestone: minor