Updated downloader to only download content
Necessary to download content only since we'll be using Shrine to attach all downloaded content as an attachment. This essentially restores the downloader removed in the previous commit with modified behavior for downloading content only. _This is a breaking change_ for firmware that will soon be fixed in future commits.

Milestone: minor