Added device migration for model and playlist foreign keys
Necessary to associate a model and/or playlist with a device. Playlists are optional so can either exist or not exist. Models are also optional but only because devices prior to applying this migration. In the future, this will need to be corrected to force a device to have a model ID.

Milestone: minor