Added problem details model payload documentation
Necessary to provide additional documentation when an API error is encountered.

This also refactors the page to enclose each problem detail in a HTML `<details>` element so when you follow the URL to this page, you'll be focused on the actual solution to the problem. Each details element has the same name so you get an accordion effect where clicking on a different problem detail will close the previous problem detail so you are not overwhelmed by all details.

Milestone: minor