Updated device schema/contract to include model and playlist IDs
Necessary to account for recently added associations (as well as prep for supporting beta firmware in the future). _This is a breaking change_, due to adding the `model_id` requirement, that will be addressed in the next commits.

This includes making more keys optional since they can be auto-generated or already have safe defaults.

Milestone: minor