Fixed device schemas and contracts to use sleep start and stop strings
Even though these are time objects, they need to come in as strings because Dry Schema will cast the incoming value to a full `Time` instance which has date information. This is not what we want. We need to use only the hour, minute, and second information which this change corrects.

Milestone: patch