Added Docker Git package
Necessary to install gems from a Git repositories. Handy for situations where you need to use a gem that isn't fully released yet. In this case, with the HTMX gem since we are pinned to a Git repository and not using a version yet (this will be corrected in the next version release of the gem).

Milestone: minor