Updated screen fetcher to answer current screen for device
Necessary to leverage a device's playlist for obtaining the current screen based on the playlist's current time. Image encryption has been removed in favor of having the screen rotator handle this isntead.

This also makes use of the sleeper to ensure rotation is ignored when the device is asleep.

Milestone: minor