Fixed RSpec device factory playlist association
Necessary to reflect default behavior since a playlist is optional and because this was creating a playlist record despite being nullified.

The repository device spec compares by attributes instead of records since ROM, when using `#combine`, will yield a struct with a `nil` value for the `playlist` key. This doesn't mimic factory behavior which seems like a bug. As a workaround, hashes are used to make this work despite being awkward.

Milestone: patch