Fixed select options view helper to use empty string for default prompt
Necessary for validation purposes, especially when dealing with integers. Example:

[source,ruby]
----
required(:playlist_id).maybe :integer
----

When using `nil`, instead of a blank string, this will cause the submitted form to translate to "Select..." which will then fail validation because we are looking for an `Integer` instead of a `String`. By using an empty string, the validation will pass because Dry Validation will translate this to a `nil` for us.

This also ensures an empty prompt is provided (single option) when no records exist.

Milestone: patch