Added model repository find by dimensions
The primary purpose for this functionality is to allow the screen synchronizer to look up model by screen resolution when syncing with the Core server. This uses a simple lookup by width to associate with our primary model. We need to update the Models API to denote which models are TRMNL specific versus community supported so this will serve as a workaround for now.

Milestone: minor