Updated designer to use name, label, and content
Necessary to ensure the new `terminus_` prefix is used for the name and that the name is what is passed between the action and the middleware to update the screen. In other words, the name *is* the ID now. The view has been updated to include the default label and also define all defaults so that the show action doesn't need to know any of this information.

Milestone: minor