Added screen gaffer
Necessary to leverage the new view, layout, and template to render an error screen for display on a device. This will create an error screen (if non-existing) or update the existing error screen if found. This is done by using `terminus_error_<device friendly ID>` as the screen name to uniquely associate the screen with a device. This also prevents multiple error screens from being rendered for a device to reduce proliferation and use a single screen for error rendering purposes.

_Gaffe is used instead of error in order to not be confused with general exceptions/errors._

Issue: 191
Milestone: minor