Refactored RSpec main application context as application dependencies
After further thought, this has been renamed to be focused on application specific dependencies without the need to use "main" in the description since the main application *is* the application. Previous iterations on this concept are here:

* 3ff75ec7c62e: Updated RSpec main application shared context to stub settings
* 2b47f76c1ec2: Refactored specs to use main application shared context

This also removes the temporary directory shared context because having the temp directory available by default is no longer necessary and can be added as needed.

Milestone: patch